requirements.txt
streamlit
google-generativeai
PyMuPDF
langchain
faiss-cpu
sentence-transformers